requirements.txt
Django==1.9.8
requests==2.24.0
pytube3==9.6.4
gunicorn